Rename packages to match directory name
