+++ lib/codeintel/repro_lang/bindings/golang/ast.go
@@ -1,4 +1,4 @@
-package repro_lang
+package golang
 
 import (
 	"fmt"
