+++ lib/codeintel/repro_lang/bindings/golang/lsif.go
@@ -1,4 +1,4 @@
-package repro_lang
+package golang
 
 import (
 	"sort"
