+++ lib/codeintel/repro_lang/bindings/golang/source_file.go
@@ -1,4 +1,4 @@
-package repro_lang
+package golang
 
 import (
 	sitter "github.com/smacker/go-tree-sitter"
