+++ lib/codeintel/repro_lang/src/binding_test.go
@@ -1,4 +1,4 @@
-package repro_lang
+package src
 
 import (
 	"context"
